# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Platform Wallet Point Gamifikasi Kampus - Backend API",
        "version": "1.0",
        "title": "Wallet Point API",
        "contact": {
            "name": "API Support",
            "email": "support@campus.edu"
        }
    },
    "host": "localhost:8102",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter your bearer token in the format: Bearer <token>"
        }
    },
    "paths": {
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Get All Users",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "full_name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "role": {
                                    "type": "string",
                                    "enum": [
                                        "admin",
                                        "dosen",
                                        "mahasiswa"
                                    ]
                                },
                                "nim_nip": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User Created"
                    }
                }
            }
        },
...
    }
}